requirements.txt
pandas
phonenumbers
sklearn.preprocessing